+++ src/main/java/org/mustangproject/ZUGFeRD/IZUGFeRDExportableContact.java
@@ -19,11 +19,11 @@
 package org.mustangproject.ZUGFeRD;
 
 /**
- * Mustangproject's ZUGFeRD implementation Neccessary interface for ZUGFeRD exporter Licensed under the APLv2
+ * Mustangproject's ZUGFeRD implementation neccessary interface for ZUGFeRD exporter Licensed under the APLv2
  *
  * @author jstaerk
  * @version 1.2.0
- * @date 2014-05-10
+ * dated 2014-05-10
  */
 
 
@@ -38,13 +38,33 @@ public interface IZUGFeRDExportableContact {
 		return null;
 	}
 
+	/**
+	 * customer global identification assigned by the seller
+	 *
+	 * @return customer identification
+	 */
+	default String getGlobalID() {
+		return null;
+	}
+
+	/**
+	 * customer global identification scheme
+	 *
+	 * @return customer identification
+	 */
+	default String getGlobalIDScheme() {
+		return null;
+	}
+
 
 	/**
 	 * First and last name of the recipient
 	 *
 	 * @return First and last name of the recipient
 	 */
-	String getName();
+	default String getName() {
+		return null;
+	}
 
 	default String getPhone() {
 		return null;