+++ src/main/java/org/mustangproject/ZUGFeRD/IZUGFeRDExportableTransaction.java
@@ -27,11 +27,20 @@ package org.mustangproject.ZUGFeRD;
  * */
 
 
+
+import java.math.BigDecimal;
+import java.util.Date;
 import org.mustangproject.ZUGFeRD.model.DocumentCodeTypeConstants;
 
-import java.util.Date;
-
 public interface IZUGFeRDExportableTransaction {
+
+	/**
+	 * appears in /rsm:CrossIndustryDocument/rsm:HeaderExchangedDocument/ram:Name
+	 * @return Name of document
+	 */
+	default String getDocumentName() {
+		return "RECHNUNG";
+	}
 
 	/**
 	 *
@@ -41,6 +50,7 @@ public interface IZUGFeRDExportableTransaction {
 	default String getDocumentCode() {
 		return DocumentCodeTypeConstants.INVOICE;
 	}
+
 
 	/**
 	 * Number, typically invoice number of the invoice
@@ -64,7 +74,7 @@ public interface IZUGFeRDExportableTransaction {
 
 	/**
 	 * this should be the full sender institution name, details, manager and tax registration. It is one of the few functions which may return null. e.g.
-	 * <p/>
+	 * <p>
 	 * Lieferant GmbH Lieferantenstraße 20 80333 München Deutschland Geschäftsführer: Hans Muster Handelsregisternummer: H A 123
 	 *
 	 * @return null or full sender institution name, details, manager and tax registration
@@ -115,6 +125,8 @@ public interface IZUGFeRDExportableTransaction {
 
 	/**
 	 * the creditors payment informations
+	 * 
+	 * @return an array of IZUGFeRDTradeSettlementPayment
 	 */
 	IZUGFeRDTradeSettlementPayment[] getTradeSettlementPayment();
 
@@ -236,4 +248,92 @@ public interface IZUGFeRDExportableTransaction {
 		return null;
 	}
 
+	/**
+	 * get reference number of the purchase order this invoice is based on
+	 *
+	 * @return the ID of the purchase order this document refers to
+	 */
+	default String getOrderReferenceNumber() {
+		return null;
+	}
+
+
+	/**
+	 * consignee identification (identification of the organisation the goods are shipped to [assigned by the costumer])
+	 *
+	 * @return the sender's identification
+	 */
+	default String getShipToOrganisationID() {
+		return null;
+	}
+
+
+	/**
+	 * consignee name (name of the organisation the goods are shipped to)
+	 *
+	 * @return the consignee's organisation name
+	 */
+	default String getShipToOrganisationName() {
+		return null;
+	}
+
+
+	/**
+	 * consignee street address (street of the organisation the goods are shipped to)
+	 *
+	 * @return consignee street address
+	 */
+	default String getShipToStreet() {
+		return null;
+	}
+
+
+	/**
+	 * consignee street postal code (postal code of the organisation the goods are shipped to)
+	 *
+	 * @return consignee postal code
+	 */
+	default String getShipToZIP() {
+		return null;
+	}
+
+
+	/**
+	 * consignee city (city of the organisation the goods are shipped to)
+	 *
+	 * @return the consignee's city
+	 */
+	default String getShipToLocation() {
+		return null;
+	}
+
+
+	/**
+	 * consignee two digit country code (country code of the organisation the goods are shipped to)
+	 *
+	 * @return the consignee's two character country iso code
+	 */
+	default String getShipToCountry() {
+		return null;
+	}
+
+	/**
+	 * get the ID of the BuyerOrderReferencedDocument, which sits in the ApplicableSupplyChainTradeAgreement
+	 * @return the ID of the document
+	 */
+	default String getBuyerOrderReferencedDocumentID() { return null; }
+
+	/**
+	 * get the issue timestamp of the BuyerOrderReferencedDocument, which sits in the ApplicableSupplyChainTradeAgreement
+	 * @return the IssueDateTime in format CCYY-MM-DDTHH:MM:SS
+	 */
+	default String getBuyerOrderReferencedDocumentIssueDateTime() { return null; }
+
+	/**
+	 * get the TotalPrepaidAmount located in SpecifiedTradeSettlementMonetarySummation (v1) or
+	 * SpecifiedTradeSettlementHeaderMonetarySummation (v2)
+	 * @return the total sum (incl. VAT) of prepayments, i.e. the difference between GrandTotalAmount
+	 * and DuePayableAmount
+	 */
+	default BigDecimal getTotalPrepaidAmount() { return BigDecimal.ZERO; }
 }