+++ src/main/java/org/mustangproject/toecount/FileChecker.java
@@ -18,6 +18,8 @@
  *********************************************************************** */
 package org.mustangproject.toecount;
 
+import java.math.BigDecimal;
+
 import org.mustangproject.ZUGFeRD.ZUGFeRDImporter;
 
 public class FileChecker {
@@ -52,7 +54,8 @@ public class FileChecker {
 		ZUGFeRDImporter zi = new ZUGFeRDImporter(filename);
 		try {
 			if (zi.canParse()) {
-				thisRun.incZUGFeRDCount();
+				thisRun.incZUGFeRDCount(zi.getVersion());
+				thisRun.incTotal(new BigDecimal(zi.getAmount()));
 				return true;
 			} else {
 				return false;