+++ History.md
@@ -1,10 +1,38 @@
+1.7.4
+=====
+2019-08-24
+
+- #102 XML entities for ZF2 export
+- corrected addAddtionaldata to addAdditionalFile
+- add zugferdimporter.getversion
+- add total amount in metrics
+- #132 Mustangproject does not work w/ Java 11
+
+
+1.7.3
+=====
+2019-08-01
+
+- #105 does not build in windows
+- #99 BOMs confuse parser
+- updated javadoc
+- fixed #104 nullpointerex when specifying no parameter
+- PR #112 complete profile options for v1 
+- access to TotalPrepaidAmount in ZF v1 #118
+- have unit tests for nDigitFormat #23
+
 
 1.7.2
 =====
+2019-07-08
 
 Support BuyerReference (r+w), as well as SpecifiedLegalOrganization (w) and DefinedTradeContact (w) 
 use dom4j to format output xml document
 corrected some exception logging glitches
+upgrade PDFBox to 2.0.15+
+extraction to use proper filename instead of alias #98
+NullPointerException in ZUGFeRDImporter.extractLowLevel #96
+Removed Bankleitzahl from ZF2
 
 
 1.7.1